backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
aiofiles==24.1.0
aiosmtplib==3.0.2
email-validator==2.2.0
python-slugify==8.0.4
Pillow==11.0.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
